+++ host/class/hid/usb_host_hid/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [unreleased]
+
+- Added global suspend/resume support
+
 ## 1.0.3
 - Fixed a bug with interface mismatch on EP IN transfer complete while several HID devices are present.
 - Fixed a bug during device freeing, while detaching one of several attached HID devices.